+++ tests/unit/vcsApp.spec.ts
@@ -1,3 +1,5 @@
+import { expect } from 'chai';
+import sinon, { type SinonSpy } from 'sinon';
 import VectorLayer from '../../src/layer/vectorLayer.js';
 import Viewpoint from '../../src/util/viewpoint.js';
 import VcsModule from '../../src/vcsModule.js';
@@ -7,12 +9,10 @@ import VcsApp from '../../src/vcsApp.js';
 describe('vcsApp', () => {
   describe('adding of a module', () => {
     describe('normal', () => {
-      let module;
-      /** @type {VcsApp} */
-      let app;
-      /** @type {Viewpoint} */
-      let startingVp;
-      let added;
+      let module: VcsModule;
+      let app: VcsApp;
+      let startingVp: Viewpoint;
+      let added: SinonSpy;
 
       before(async () => {
         startingVp = new Viewpoint({
@@ -59,7 +59,7 @@ describe('vcsApp', () => {
       it('should load layers which are active on startup', () => {
         const layer = app.layers.getByKey('bar');
         expect(layer).to.be.an.instanceOf(VectorLayer);
-        expect(layer.active || layer.loading).to.be.true; // we do not wait for layers. so all good _as long as its not inactive_
+        expect(layer!.active || layer!.loading).to.be.true; // we do not wait for layers. so all good _as long as its not inactive_
       });
 
       it('should activate the starting map', () => {
@@ -67,24 +67,22 @@ describe('vcsApp', () => {
       });
 
       it('should activate the starting view point', () => {
-        const vp = app.maps.activeMap.getViewpointSync();
-        expect(vp.equals(startingVp, 10e-8)).to.be.true;
+        const vp = app.maps.activeMap?.getViewpointSync();
+        expect(vp?.equals(startingVp, 10e-8)).to.be.true;
       });
 
       it('should globaly hide hidden objects', () => {
         expect(app.layers.globalHider.hiddenObjects).to.have.keys(
-          module.config.hiddenObjects.map((o) => o.id),
+          module.config.hiddenObjects!.map((o) => o.id),
         );
       });
     });
 
     describe('adding of a second module', () => {
-      let module;
-      /** @type {VcsApp} */
-      let app;
-      /** @type {Viewpoint} */
-      let startingVp;
-      let added;
+      let module: VcsModule;
+      let app: VcsApp;
+      let startingVp: Viewpoint;
+      let added: SinonSpy;
 
       before(async () => {
         startingVp = new Viewpoint({
@@ -115,7 +113,7 @@ describe('vcsApp', () => {
         added = sinon.spy();
         app.moduleAdded.addEventListener(added);
         await app.addModule(initialModule);
-        app.layers.getByKey('bar').deactivate();
+        app.layers.getByKey('bar')!.deactivate();
         await app.addModule(module);
       });
 
@@ -139,13 +137,13 @@ describe('vcsApp', () => {
       it('should load layers which are active on startup', () => {
         const layer = app.layers.getByKey('foo');
         expect(layer).to.be.an.instanceOf(VectorLayer);
-        expect(layer.active || layer.loading).to.be.true; // we do not wait for layers. so all good _as long as its not inactive_
+        expect(layer!.active || layer!.loading).to.be.true; // we do not wait for layers. so all good _as long as its not inactive_
       });
 
       it('should not load layers which are active on startup and not in the current module', () => {
         const layer = app.layers.getByKey('bar');
         expect(layer).to.be.an.instanceOf(VectorLayer);
-        expect(layer.active || layer.loading).to.be.false; // we do not wait for layers. so all good _as long as its not inactive_
+        expect(layer!.active || layer!.loading).to.be.false; // we do not wait for layers. so all good _as long as its not inactive_
       });
 
       it('should only hide an object once module', () => {
@@ -155,11 +153,9 @@ describe('vcsApp', () => {
     });
 
     describe('without an active map name', () => {
-      let module;
-      /** @type {VcsApp} */
-      let app;
-      /** @type {Viewpoint} */
-      let startingVp;
+      let module: VcsModule;
+      let app: VcsApp;
+      let startingVp: Viewpoint;
 
       before(async () => {
         startingVp = new Viewpoint({
@@ -194,12 +190,10 @@ describe('vcsApp', () => {
     });
 
     describe('if activating the same module twice', () => {
-      let module;
-      /** @type {VcsApp} */
-      let app;
-      /** @type {Viewpoint} */
-      let startingVp;
-      let added;
+      let module: VcsModule;
+      let app: VcsApp;
+      let startingVp: Viewpoint;
+      let added: SinonSpy;
 
       before(async () => {
         startingVp = new Viewpoint({
@@ -220,7 +214,8 @@ describe('vcsApp', () => {
         app = new VcsApp();
         added = sinon.spy();
         app.moduleAdded.addEventListener(added);
-        app.addModule(module);
+        // eslint-disable-next-line no-void
+        void app.addModule(module);
         await app.addModule(module);
       });
 
@@ -246,7 +241,7 @@ describe('vcsApp', () => {
       it('should load layers which are active on startup', () => {
         const layer = app.layers.getByKey('bar');
         expect(layer).to.be.an.instanceOf(VectorLayer);
-        expect(layer.active || layer.loading).to.be.true; // we do not wait for layers. so all good _as long as its not inactive_
+        expect(layer!.active || layer!.loading).to.be.true; // we do not wait for layers. so all good _as long as its not inactive_
       });
 
       it('should activate the starting map', () => {
@@ -254,20 +249,84 @@ describe('vcsApp', () => {
       });
 
       it('should activate the starting view point', () => {
-        const vp = app.maps.activeMap.getViewpointSync();
-        expect(vp.equals(startingVp, 10e-8)).to.be.true;
+        const vp = app.maps.activeMap?.getViewpointSync();
+        expect(vp?.equals(startingVp, 10e-8)).to.be.true;
+      });
+    });
+
+    describe('if activating a module with the same _id', () => {
+      let module: VcsModule;
+      let app: VcsApp;
+      let startingVp: Viewpoint;
+      let added: SinonSpy;
+
+      before(async () => {
+        startingVp = new Viewpoint({
+          name: 'foo',
+          groundPosition: [13, 52],
+          distance: 200,
+        });
+        module = new VcsModule({
+          _id: 'foo',
+          layers: [
+            new VectorLayer({ name: 'foo' }).toJSON(),
+            new VectorLayer({ name: 'bar', activeOnStartup: true }).toJSON(),
+          ],
+          viewpoints: [new Viewpoint({}).toJSON(), startingVp.toJSON()],
+          maps: [new OpenlayersMap({ name: 'foo' }).toJSON()],
+          startingViewpointName: 'foo',
+          startingMapName: 'foo',
+        });
+        app = new VcsApp();
+        added = sinon.spy();
+        app.moduleAdded.addEventListener(added);
+        // eslint-disable-next-line no-void
+        void app.addModule(module);
+        await app.addModule(new VcsModule(module.config));
+      });
+
+      after(() => {
+        app.destroy();
+      });
+
+      it('should add the module once', () => {
+        expect(app.getModuleById(module._id)).to.equal(module);
+      });
+
+      it('should raise the moduleAdded event once', () => {
+        expect(added).to.have.been.calledOnce;
+      });
+
+      it('should add the modules resources', () => {
+        expect(app.layers.hasKey('foo')).to.be.true;
+        expect(app.layers.hasKey('bar')).to.be.true;
+        expect(app.maps.hasKey('foo')).to.be.true;
+        expect(app.viewpoints.hasKey('foo')).to.be.true;
+      });
+
+      it('should load layers which are active on startup', () => {
+        const layer = app.layers.getByKey('bar');
+        expect(layer).to.be.an.instanceOf(VectorLayer);
+        expect(layer!.active || layer!.loading).to.be.true; // we do not wait for layers. so all good _as long as its not inactive_
+      });
+
+      it('should activate the starting map', () => {
+        expect(app.maps.activeMap).to.have.property('name', 'foo');
+      });
+
+      it('should activate the starting view point', () => {
+        const vp = app.maps.activeMap?.getViewpointSync();
+        expect(vp?.equals(startingVp, 10e-8)).to.be.true;
       });
     });
   });
 
   describe('removing a module', () => {
     describe('normal', () => {
-      let module;
-      /** @type {VcsApp} */
-      let app;
-      /** @type {Viewpoint} */
-      let startingVp;
-      let removed;
+      let module: VcsModule;
+      let app: VcsApp;
+      let startingVp: Viewpoint;
+      let removed: SinonSpy;
 
       before(async () => {
         startingVp = new Viewpoint({
@@ -313,18 +372,16 @@ describe('vcsApp', () => {
 
       it('should show globaly hidden objects', () => {
         expect(app.layers.globalHider.hiddenObjects).to.not.have.keys(
-          module.config.hiddenObjects.map((o) => o.id),
+          module.config.hiddenObjects!.map((o) => o.id),
         );
       });
     });
 
     describe('if removing the same module twice', () => {
-      let module;
-      /** @type {VcsApp} */
-      let app;
-      /** @type {Viewpoint} */
-      let startingVp;
-      let removed;
+      let module: VcsModule;
+      let app: VcsApp;
+      let startingVp: Viewpoint;
+      let removed: SinonSpy;
 
       before(async () => {
         startingVp = new Viewpoint({
@@ -370,8 +427,7 @@ describe('vcsApp', () => {
   });
 
   describe('locale', () => {
-    /** @type {VcsApp} */
-    let app;
+    let app: VcsApp;
 
     before(() => {
       app = new VcsApp();